restructures for testing framework
-failures in testing from incorrect imports / file structure
