+++ src/vardict/vardict_class.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 # imports
 import os
 import sys
